fix(wallet-sdk): improve file structure
