--- packages/libs/wallet-sdk/package.cjs.json
@@ -1,3 +1,0 @@
-{
-  "type": "commonjs"
-}
