+++ packages/libs/wallet-sdk/src/sdk/kadenaNames.ts
@@ -1,0 +1,54 @@
+import {
+  addressToName,
+  nameToAddress,
+} from '../services/kadenaNamesService.js';
+import type { WalletSDK } from './walletSdk.js';
+
+export class KadenaNames {
+  private _sdk: WalletSDK;
+
+  public constructor(walletSDK: WalletSDK) {
+    this._sdk = walletSDK;
+  }
+
+  public async nameToAddress(
+    name: string,
+    networkId: string,
+  ): Promise<string | null> {
+    try {
+      const host = this._sdk.hostUrlGenerator({
+        networkId,
+        chainId: '15',
+      });
+      const result = await nameToAddress(name, networkId, host);
+
+      if (result === undefined) {
+        console.warn(`No address found for name: ${name}`);
+        return null;
+      }
+      return result;
+    } catch (error) {
+      console.error(`Error in name resolving action: ${error.message}`);
+      throw new Error(`Error resolving address: ${error.message}`);
+    }
+  }
+
+  public async addressToName(
+    address: string,
+    networkId: string,
+  ): Promise<string | null> {
+    try {
+      const host = this._sdk.hostUrlGenerator({ networkId, chainId: '15' });
+      const result = await addressToName(address, networkId, host);
+
+      if (result === undefined) {
+        console.warn(`No address found for name: ${address}`);
+        return null;
+      }
+      return result;
+    } catch (error) {
+      console.error(`Error in name resolving action: ${error.message}`);
+      throw new Error(`Error resolving address: ${error.message}`);
+    }
+  }
+}
