+++ packages/libs/wallet-sdk/src/sdk/tests/walletSdk.test.ts
@@ -1,0 +1,7 @@
+import { assert, describe, test } from 'vitest';
+
+describe('example test', () => {
+  test('example', async () => {
+    assert(true);
+  });
+});
