+++ packages/libs/wallet-sdk/src/services/tests/accountService.ts
@@ -1,0 +1,7 @@
+import { assert, describe, test } from 'vitest';
+
+describe('example test', () => {
+  test('example', async () => {
+    assert(true);
+  });
+});
